# Partial expand of casts to non-sealed classes