Currently, we don't expand casts/isinst to non-sealed classes, e.g.
```csharp
public class MyClass
{
    static bool Test(object o) => o is MyClass;


    static void Test2(object o)
    {
        if (o is MyClass p)
        {
            p.DoWork();
        }
    }


    [MethodImpl(MethodImplOptions.NoInlining)]
    public void DoWork() {}
}
```
Here is the codegen for `Test` and `Test2`: 
```asm
; Method Program:Test(System.Object):bool
G_M4425_IG01:
       sub      rsp, 40
G_M4425_IG02:
       mov      rdx, rcx
       mov      rcx, 0xD1FFAB1E
       call     CORINFO_HELP_ISINSTANCEOFCLASS
       test     rax, rax
       setne    al
       movzx    rax, al
G_M4425_IG03:
       add      rsp, 40
       ret      
; Total bytes of code: 36


; Method Program:Test2(System.Object)
G_M61455_IG01:
       sub      rsp, 40
G_M61455_IG02:
       mov      rdx, rcx
       mov      rcx, 0xD1FFAB1E
       call     CORINFO_HELP_ISINSTANCEOFCLASS
       test     rax, rax
       je       SHORT G_M61455_IG05
G_M61455_IG03:
       mov      rcx, rax
G_M61455_IG04:
       add      rsp, 40
       jmp      Program:DoWork():this
G_M61455_IG05:
       add      rsp, 40
       ret      
; Total bytes of code: 44
```
As you can see, we just call `CORINFO_HELP_ISINSTANCEOFCLASS`.
I propose we also expand cases like this but with that call as a fallback. 
Basically, optimize `if (CORINFO_HELP_ISINSTANCEOFCLASS(obj, MyNonSealedType)` to `if ((obj != null && obj->pMT == MyNonSealedType) || CORINFO_HELP_ISINSTANCEOFCLASS(obj, MyNonSealedType))`


Here is the new codegen:
```asm
; Method Program:Test(System.Object):bool
G_M4425_IG01:
       sub      rsp, 40
G_M4425_IG02:
       mov      gword ptr [rsp+30H], rcx
       mov      rax, rcx
       test     rax, rax  ;; o is null ?
       je       SHORT G_M4425_IG05
G_M4425_IG03:
       mov      rdx, 0xD1FFAB1E
       cmp      qword ptr [rax], rdx  ;; o is exactly BaseClass ?
       je       SHORT G_M4425_IG05
G_M4425_IG04:
       mov      rcx, rdx
       mov      rdx, gword ptr [rsp+30H]
       call     CORINFO_HELP_ISINSTANCEOFCLASS ;; it wasn't BaseClass - fallback to helper call
G_M4425_IG05:
       test     rax, rax ;; redundant, but it's a different issue.
       setne    al
       movzx    rax, al
G_M4425_IG06:
       add      rsp, 40
       ret      
; Total bytes of code: 59


; Method Program:Test2(System.Object)
G_M61455_IG01:
       sub      rsp, 40
G_M61455_IG02:
       mov      gword ptr [rsp+30H], rcx
       mov      rax, rcx
       test     rax, rax  ;; o is null ?
       je       SHORT G_M61455_IG05
G_M61455_IG03:
       mov      rdx, 0xD1FFAB1E
       cmp      qword ptr [rax], rdx   ;; o is exactly BaseClass ?
       je       SHORT G_M61455_IG05
G_M61455_IG04:
       mov      rcx, rdx
       mov      rdx, gword ptr [rsp+30H]
       call     CORINFO_HELP_ISINSTANCEOFCLASS ;; it wasn't BaseClass - fallback to helper call
G_M61455_IG05:
       test     rax, rax ;; redundant, but it's a different issue.
       je       SHORT G_M61455_IG08
G_M61455_IG06:
       mov      rcx, rax
G_M61455_IG07:
       add      rsp, 40
       jmp      Program:DoWork():this
G_M61455_IG08:
       add      rsp, 40
       ret      
; Total bytes of code: 67
```
[diffchecker](https://www.diffchecker.com/WWkrg8j1).

### Benchmarks
I wrote a simple benchmark to check benefits or overhead (in case of subclasses): https://gist.github.com/EgorBo/a3ac2b1f5c6e182acdd5f5d43c8335ec

Here are the results:
```
     |      Method |         o |      Mean |     Error |    StdDev | Ratio |
     |------------ |---------- |----------:|----------:|----------:|------:|
  PR | IsBaseClass |      null | 0.2856 ns | 0.0168 ns | 0.0157 ns |  0.20 |
Main | IsBaseClass |      null | 1.4008 ns | 0.0109 ns | 0.0091 ns |  1.00 |
     |             |           |           |           |           |       |
  PR |  CallDoWork |      null | 0.5540 ns | 0.0129 ns | 0.0114 ns |  0.29 |
Main |  CallDoWork |      null | 1.9152 ns | 0.0155 ns | 0.0129 ns |  1.00 |
     |             |           |           |           |           |       |
  PR | IsBaseClass | BaseClass | 0.3257 ns | 0.0240 ns | 0.0224 ns |  0.21 |
Main | IsBaseClass | BaseClass | 1.5435 ns | 0.0228 ns | 0.0214 ns |  1.00 |
     |             |           |           |           |           |       |
  PR |  CallDoWork | BaseClass | 1.0866 ns | 0.0234 ns | 0.0207 ns |  0.56 |
Main |  CallDoWork | BaseClass | 1.9415 ns | 0.0153 ns | 0.0136 ns |  1.00 |
     |             |           |           |           |           |       |
  PR | IsBaseClass |   Program | 2.0706 ns | 0.0213 ns | 0.0189 ns |  0.98 |
Main | IsBaseClass |   Program | 2.1222 ns | 0.0398 ns | 0.0353 ns |  1.00 |
     |             |           |           |           |           |       |
  PR |  CallDoWork |   Program | 2.5887 ns | 0.0168 ns | 0.0140 ns |  0.88 |
Main |  CallDoWork |   Program | 2.9370 ns | 0.0469 ns | 0.0439 ns |  1.00 |
     |             |           |           |           |           |       |
  PR | IsBaseClass |  Subclass | 2.2969 ns | 0.0136 ns | 0.0114 ns |  1.10 |
Main | IsBaseClass |  Subclass | 2.0966 ns | 0.0369 ns | 0.0345 ns |  1.00 |
     |             |           |           |           |           |       |
  PR |  CallDoWork |  Subclass | 3.0225 ns | 0.0204 ns | 0.0191 ns |  1.00 |
Main |  CallDoWork |  Subclass | 3.0209 ns | 0.0194 ns | 0.0182 ns |  1.00 |
```
A 10% (+0.2ns) overhead for a case when `Subclass` is passed (and we expected `BaseClass`), but a significant win for cases where `null` or `BaseClass` are passed.

PS: we'll avoid the overhead at all with PGO, see https://github.com/dotnet/runtime/issues/55325
PS2: the new codegen is not super-efficient and will be improved, see https://github.com/dotnet/runtime/issues/55841#issuecomment-881589691)