@EgorBo 2021-07-18
The diff is quote big so I guess I can enable this opt only for really hot methods (with PGO data).