@EgorBo 2021-07-19
Closing in favor of a proper fix for https://github.com/dotnet/runtime/issues/55325